Text Channel (#22)
* react component

* Create TextChannel.css

* Update App.js

removed html and javascript code that are for a component
